+++ modules/optics/arrow-optics/src/main/kotlin/arrow/optics/std/validated.kt
@@ -2,10 +2,8 @@ package arrow.optics
 
 import arrow.core.*
 import arrow.data.*
-import arrow.data.Failure
-import arrow.data.Success
-import arrow.optics.Iso
-import arrow.optics.PIso
+import arrow.core.Failure
+import arrow.core.Success
 
 /**
  * [PIso] that defines equality between [Validated] and [Either]
