+++ modules/optics/arrow-optics/src/test/kotlin/arrow/optics/instances/EachInstanceTest.kt
@@ -5,7 +5,7 @@ import arrow.core.EitherPartialOf
 import arrow.core.Option
 import arrow.data.ListK
 import arrow.data.MapK
-import arrow.data.Try
+import arrow.core.Try
 import arrow.optics.typeclasses.each
 import arrow.test.UnitSpec
 import arrow.test.generators.genEither
